+++ client/src/lvgl-gui/gui_common.h
@@ -15,7 +15,7 @@ typedef enum {
 
 typedef struct {
   void *card;
-  uint8_t prev_view;
+  uint16_t prev_view;
 } CardData;
 
 iso14a_card_select_t *hf14a_read(bool keep_field, bool skip_RATS, bool ecp,
